refactor contract
