+++ contracts/engagement/src/core/admin.rs
@@ -1,6 +1,6 @@
 use soroban_sdk::{Address, Env};
 
-use crate::storage_types::DataKey;
+use crate::storage::types::DataKey;
 
 pub fn has_administrator(e: &Env) -> bool {
     let key = DataKey::Admin;
