+++ contracts/engagement/src/events/mod.rs
@@ -1,0 +1,7 @@
+pub mod handlers;
+
+pub (crate) use handlers::{
+    escrows_by_engagement_id,
+    balance_retrieved_event,
+    allowance_retrieved_event
+};
