+++ contracts/engagement/src/storage/mod.rs
@@ -1,0 +1,2 @@
+pub mod types;
+mod persistence;
