+++ contracts/engagement/src/storage/persistence.rs
