--- contracts/engagement/src/utils.rs
@@ -1,3 +1,0 @@
-use soroban_sdk::{BytesN, Env, FromVal, String};
-
-// Here will go future functions
